Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main(){
-    int x,y,z;
+    float x,y,z;
     scanf("%d",&x);
     scanf("%d",&y);
     scanf("%d",&z);
